+++ code/utils/utils.hpp
@@ -5,7 +5,7 @@
 // ############################## VERSION ############################## //
 // ##################################################################### //
 
-#define VERSION "2.2.2:4"
+#define VERSION "2.2.2:5"
 
 // ##################################################################### //
 // ############################## IMPORTS ############################## //
@@ -45,6 +45,13 @@
 #define HPLUS_CLI_FRACTCUTS_FLAG "fract"
 #define HPLUS_CLI_CANDCUTS_FLAG "cand"
 #define HPLUS_CLI_CUTLOOP_FLAG "cloop"
+#define HPLUS_CLI_CUTLOOP_MIN_ITER_FLAG "cl-min-iter"
+#define HPLUS_CLI_CUTLOOP_IMPROVEMENT_FLAG "cl-improv"
+#define HPLUS_CLI_CUTLOOP_PAST_ITER_FLAG "cl-past-iter"
+#define HPLUS_CLI_INOUT_FLAG "inout"
+#define HPLUS_CLI_INOUT_MAX_ITER_FLAG "io-max-it"
+#define HPLUS_CLI_INOUT_WEIGHT_FLAG "io-w"
+#define HPLUS_CLI_INOUT_WEIGHT_UPD_FLAG "io-wupd"
 #define HPLUS_CLI_TESTING_FLAG "test"
 
 // ##################################################################### //
@@ -73,7 +80,14 @@
 #define HPLUS_DEF_VERBOSE 3
 #define HPLUS_DEF_FRACTCUTS "ls"
 #define HPLUS_DEF_CANDCUTS "fcs"
-#define HPLUS_DEF_CUSTOM_CUTLOOP 0
+#define HPLUS_DEF_CUSTOM_CUTLOOP false
+#define HPLUS_DEF_CL_MIN_ITER 20
+#define HPLUS_DEF_CL_IMPROV 0.005
+#define HPLUS_DEF_CL_PAST_ITER 10
+#define HPLUS_DEF_INOUT true
+#define HPLUS_DEF_IO_MAX_IT 4
+#define HPLUS_DEF_IO_WEIGHT .4
+#define HPLUS_DEF_IO_WEIGHT_UPD .5
 
 // ##################################################################### //
 // ######################### EXECUTION DEFAULTS ######################## //